PEH-3281 - Modal Window Icon and Date Error Display
On the home page - Study Visit Details > when there is a link in the Visit Column, it opens a modal window. The CSS that was being applied to the tile as also being applied to the display in the modal window. The Icons were limited to 5 and an ellipsis was showing. The specific display should be limited to just the Study Visit Details Tile and the modal should display all icons, wrapping at 60%.

The Date and Time input error display below the fields were overlapping - the error text extending past the width of the input field. The text should wrap and there needs to be sufficient space between the error and the label below.