PEH-3342
Had to modify spacing between elements to compensate for error messages and a rendering delay when an error shows. The modal would open, spacing was smaller than it should be, an error triggered with error display. and a mouse click anywhere would suddenly add more space. This is the addition of a lightning error div that has no content but takes up space. Fixed spacing vertically between checkboxes. Fixed alignment of travel title and travel button. Both needed padding removed from sides and button was not aligned to the right per mockup.