REF-3855 - Red Mandatory Error for Attestation checkbox
All (hopefully) instances of the "Complete this field" for the Year of Birth checkbox have been addressed to have it not wrap. For the different screen sizes, the top margin was adjusted to move it down below the label since the label wraps to more lines. 

On the Patient Details > modal > Sharing Options tab, the "Connect to Patient" button is located to the right of the Last Name input field. When the fields are filled in, the Year of Birth fields are added which means the button is above the last input. A CSS Order was added so that the button is the last in the area meaning when the YOB is not visible, the button is on the right; when the YOB appears, the button will appear below the checkbox for ALL screens. 

In Pre-Screening Eligibility, the error message was moved down for all screens but I noticed that the page WAS NOT set up for responsiveness. I changed all the fields areas to 100% and widened the YOB fields so that the year is completely visible (it was cut off). I also moved the page content down, the CONGRATULATIONS text was right up to the top of the area (added 25px padding inline) - I thought I added this weeks ago(?). The "Participant is No Longer Interested" button was left aligned and should be centered. 